Exercise 04 - things to note
Initially used splice to find and remove indices. However, the
splice method reduced the length of the array, which was affecting
the loop variable. Hence it was not getting to the end. I added
a 'minus i' line IF a value was removed so that I wouldn't skip,
but it ended in an endless loop. Decided to just create a new
list that appended items only if they weren't matching the inputs.

There, proper commit done and dusted! Onwards.